final fixes for Windows build

+++ createdbc.cpp
@@ -64,7 +64,7 @@ int main (int argc, char* argv[])
 	for (int ii=1; ii < argc; ++ii)
 	{
 		if (kStrIn (argv[ii], "-d,-dd,-ddd")) {
-			KLog().SetLevel( strlen(argv[ii]) - 1 );
+			KLog().SetLevel( static_cast<int>(strlen(argv[ii])) - 1 );
 			KLog().SetDebugLog("stdout");
 			kDebugLog (0, "{}: debug now set to {}", argv[ii], KLog().GetLevel());
 		}
